requirements.txt
streamlit
sqlalchemy
pytesseract
tensorflow
pillow
opencv-python
pandas